app/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["Dert2 <dert.0.1@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
databases = {extras = ["asyncpg"], version = "^0.6.1"}
python-dotenv = "^0.21.0"
aiosqlite = "^0.17.0"
psycopg2-binary = "^2.9.4"
validators = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
